refactor: separate renderer
